hrms/requirements.txt
asgiref==3.9.1
click==8.2.1
cmake==4.1.0
colorama==0.4.6
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
dlib==20.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==2.2.6
pillow==11.3.0
psycopg2==2.9.10
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
